# user-module/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User module API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/history/users/{id}": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Get detail history user",
                "operationId": "ee86d626724983c2708b0e32ebb54bdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of history user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail history user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "History"
                ],
                "summary": "Restore user",
                "operationId": "58f15103f8b53f5b0ff7520a59bb1103",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of history user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restore user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/history/departments/{id}": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Get detail history department",
                "operationId": "0ef4d25062f275e7c341770c6cfda4b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of history department",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail history department"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "History"
                ],
                "summary": "Restore department",
                "operationId": "a613327ca5d593e254df691aba537e0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of history department",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restore department"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/history/users": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Get list history users",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of user",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort by column",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List history users"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/history/departments": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Get list history departments",
                "parameters": [
                    {
                        "name": "department_id",
                        "in": "query",
                        "description": "Id of department",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort by column",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List history departments"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/history/notifications": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Get list history notifications",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of history notification",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort by column",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List history notifications"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}